Refactor test utilities and improve test assertions in project tests
