--- test/project/project.test.ts
@@ -1,8 +1,0 @@
-/**
- * Project class tests are organized into separate files for better maintainability:
- *
- * - project-package.test.ts: Tests for package operations (unpack, create, update)
- * - project-dependencies.test.ts: Tests for dependency management (install, add, remove)
- *
- * See those files for the actual test implementations.
- */
